host-custom/Asterisk_Integration/requirements.txt
asterisk-ami
requests

